youtube_downloader/requirements.txt
pytube
mysql-connector-python
google-cloud-firestore
python-dotenv
ttkbootstrap